Allow `main()` to be called with zero arguments
Standard packaging tools such as pip or PyPA’s `installer` use a feature
called entry points, so they can automatically generate platform
executables on the fly [1] [2] as the user installs a wheel. That allows
users to type e.g. `shrinko8` instead of `python shrinko8.py`.

Entry points are functions that offer a zero-arguments form. [3] [4]
Contrary to languages such as C or Java, an entry point in Python
doesn’t conveniently receive command-line arguments via function
parameters. Instead, it is expected to actively query the arguments from
e.g. `sys.argv` if need be.

Adding a zero-argument form to `main` is a first step towards making
shrinko8 compatible with standard packaging tools, and helps in case you
ever decide to publish shrinko8 on PyPI and allow users to `pip install`
it.

It also makes it a little easier for system-level package maintainers to
package `shrinko8` and `shrinkotron` as executables.

[1]: https://github.com/pypa/pip/blob/06c8024bc50b198b91cad6df80d72a742d524e3d/src/pip/_vendor/distlib/scripts.py#L42-L49

[2]: https://github.com/pypa/installer/blob/7656f5d41943c21757243efa9deef636591bada2/src/installer/scripts.py#L35-L43

[3]: https://discuss.python.org/t/why-do-script-entrypoints-require-a-function-be-specified/14090

[4]: https://packaging.python.org/en/latest/specifications/entry-points/